Inputs: move, halt, jump
In this commit, we use Android touch inputs to distinguish three
different commands, which the loop will use to control the player
character.

A move happens whenever the user presses down on any part of the screen,
until he lifts the finger back up.

A halt is simply the end of the touch gesture. It means to stop moving.

A jump is achieved either through a second-finger tap, or a double-tap
with the primary finger anywhere on the screen.

The game loop does not yet apply these inputs to the ingame objects. It
merely outputs the commands received to the debug log.