Fullscreen fixes
We now hide all navigation elements when we enter the GameActivity.
This should probably be adapted for all other activities later.

All touch coordinates are now properly converted to game board
coordinates.

Some debug solid lines illustrate the edge of the game board on the
screen, which should give us a feeling for the accuracy of the
coordinate conversion.